:bug: Fixed random generator/seed mechanism
There was a bug when the passed seed was a np.int32 instead of an integer. Functions now check is the passed generator/seed is instance of numpy.random.Generator instead of checking if it is an instance of int. The function random_nan_binariser was enhanced with a 10x speedup and optional parameters.